// Food-Delivery-WebApp_Backend/src/models/Food.ts
import mongoose, { Schema, model } from "mongoose";

export interface Food{
    id:string;
    name:string;
    price:number;
    tags: string[];
    favorite:boolean;
    stars: number;
    imageUrl: string;
    origins: string[];
    cookTime:string;

}

export const FoodSchema=new Schema<Food>({




    
    name:{type:String,required:true},
    price:{type:Number,required:true},
    tags: {type:[String]},
    favorite:{type:Boolean,default:true},
    stars: {type:Number,required:true},
    imageUrl: {type:String,required:true},
    origins:  {type:[String],required:true},
    cookTime:{type:String,required:true},

}
,

{
    // toJSON is required to map your id field with the _id field in the doccment of db , as its an autogenerated id
    toJSON: { virtuals: true },    // Include virtual properties when converting to JSON
    toObject: { virtuals: true },  // Include virtual properties when converting to plain objects
    timestamps: true              // Automatically manage createdAt and updatedAt fields
}
)

export const FoodModel=model<Food>('food', FoodSchema)